Refactor all filters to one component
