+++ src/components/inventory/InventoryFilter.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const InventoryFilter = () => {
+  return (
+    <div>InventoryFilter</div>
+  )
+}
+
+export default InventoryFilter
